+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/receiptschedule/impl/HUReceiptScheduleWeightNetAdjuster.java
@@ -286,7 +286,6 @@ public class HUReceiptScheduleWeightNetAdjuster
 
 		//
 		// Case: WeightNet is same as HU Storage Qty
-		// => do thing, this is what we try to achieve
 		if (qtyToAllocateAbs.signum() == 0)
 		{
 			logger.debug("HU Storage's Qty is same as WeightNet => nothing to do");
@@ -302,8 +301,8 @@ public class HUReceiptScheduleWeightNetAdjuster
 		// =>Transfer from Receipt schedule to HU Storage
 		if (qtyToAllocateAbs.signum() > 0)
 		{
-			source = createAllocationSourceDestination(receiptSchedule);
-			destination = createAllocationDestination(vhu);
+			source = createSourceOrDestForReceiptSchedule(receiptSchedule);
+			destination = createSourceOrDestForVHU(vhu);
 			qtyToAllocate = qtyToAllocateAbs;
 		}
 		//
@@ -311,8 +310,8 @@ public class HUReceiptScheduleWeightNetAdjuster
 		// =>Transfer from HU Storage to Receipt schedule
 		else
 		{
-			source = createAllocationSource(vhu);
-			destination = createAllocationSourceDestination(receiptSchedule);
+			source = createSourceOrDestForVHU(vhu);
+			destination = createSourceOrDestForReceiptSchedule(receiptSchedule);
 			qtyToAllocate = qtyToAllocateAbs.negate();
 		}
 
@@ -348,21 +347,20 @@ public class HUReceiptScheduleWeightNetAdjuster
 		}
 	}
 
-	private AbstractAllocationSourceDestination createAllocationSourceDestination(final I_M_ReceiptSchedule receiptSchedule)
+	private HUListAllocationSourceDestination createSourceOrDestForVHU(final I_M_HU vhu)
+	{
+		final HUListAllocationSourceDestination huListAllocationSourceDestination = new HUListAllocationSourceDestination(vhu);
+		
+		// gh #943: in case of aggregate HUs we *only* want to change the storage. we don't want AggregateHUTrxListener do change the TU-per-CU quantity or do other stuff.
+		huListAllocationSourceDestination.setStoreCUQtyBeforeProcessing(false);
+		return huListAllocationSourceDestination;
+	}
+
+	private AbstractAllocationSourceDestination createSourceOrDestForReceiptSchedule(final I_M_ReceiptSchedule receiptSchedule)
 	{
 		final IProductStorage storage = huReceiptScheduleBL.createProductStorage(receiptSchedule);
 		final Object trxReferencedModel = receiptSchedule;
 		return new GenericAllocationSourceDestination(storage, trxReferencedModel);
-	}
-
-	private IAllocationSource createAllocationSource(final I_M_HU hu)
-	{
-		return new HUListAllocationSourceDestination(hu);
-	}
-
-	private IAllocationDestination createAllocationDestination(final I_M_HU hu)
-	{
-		return new HUListAllocationSourceDestination(hu);
 	}
 
 	private boolean isEligible(final I_M_ReceiptSchedule_Alloc rsa)